refactor: rename costcalc to chatcost and squash migrations
